move mappreview extra creation into the create-or-update-dataset process; allow for passing layers to display via config (#164)
