+++ src/ckanext-natcap/ckanext/natcap/templates/package/edit_base.html
@@ -1,6 +1,5 @@
 {% ckan_extends %}
 
 {% block content_action %}
-    {% snippet "package/snippets/update_mappreview.html", pkg=pkg %}
     {{ super() }}
 {% endblock %}
