+++ src/ckanext-natcap/ckanext/natcap/templates/package/read_base.html
@@ -1,7 +1,6 @@
 {% ckan_extends %}
 
 {% block content_action %}
-    {% snippet "package/snippets/update_mappreview.html", pkg=pkg %}
     {{ super() }}
 {% endblock %}
